Move Order Command into order package
